extract BufferStandardDescriptors to common_test
